ReImproved Casting(Implicit)

+++ Variable/Three_Variable_TypeCasting/Program.cs
@@ -1,7 +1,7 @@
 ﻿// See https://aka.ms/new-console-template for more information
 Console.WriteLine("Hello, World!");
 byte smallNumber = 100;
-int largeNumber = (int)smallNumber;
+int largeNumber = smallNumber;
 Console.WriteLine(largeNumber);//Implicit Casting
 double largerNumber = 300.45;
 int smallerNumber = (int)largerNumber;//Explicit Casting //Type Casting//Data may be lost
